Final solution - code could use some cleanup

+++ Dijkstra/Dijkstra/large.h
@@ -6,4 +6,4 @@
 #include <vector>
 #include "common.h"
 
-bool SolveLargeProblem(long L, long X, std::string line);
+bool SolveLargeProblem(long long L, long long X, std::string line);
